pyproject.toml
[project]
name = "sample-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.12.0",
    "imgcat>=0.6.0",
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.2",
    "langgraph-checkpoint>=2.1.1",
    "langsmith>=0.4.13",
    "markdown>=3.8.2",
    "matplotlib>=3.10.5",
    "mcp[cli]>=1.12.4",
    "pillow>=11.3.0",
    "pydantic==2.10.6",
    "pygraphviz>=1.14",
    "retry>=0.9.2",
    "yfinance>=0.2.65",
]
